+++ CHANGES.rst
@@ -22,6 +22,7 @@ Minor Changes
 - fix(concurrency): rely on context manager boolean control flow
 - docs(coc): fix typos & rename first section
 - docs(perm): add missing operator in doc string examples
+- docs(cipher): salt misuse resistance calculations inaccurate [#16]
 - docs(concurrency): fix typos & grammar in doc strings
 - style(lint): turn on "ARG" flag & fix associated warnings